+++ atlite/gis.py
@@ -26,7 +26,7 @@ from shapely.ops import transform
 from rasterio.warp import reproject, transform_bounds
 from rasterio.mask import mask
 from rasterio.features import geometry_mask
-from scipy.ndimage.morphology import binary_dilation as dilation
+from scipy.ndimage import binary_dilation as dilation
 from numpy import isin, empty
 from shapely.strtree import STRtree
 from tqdm import tqdm
@@ -141,12 +141,15 @@ def compute_indicatormatrix(orig, dest, orig_crs=4326, dest_crs=4326):
     dest = reproject_shapes(dest, dest_crs, orig_crs)
     indicator = sp.sparse.lil_matrix((len(dest), len(orig)), dtype=float)
     tree = STRtree(orig)
-    idx = dict((id(o), i) for i, o in enumerate(orig))
+    idx = dict((hash(o.wkt), i) for i, o in enumerate(orig))
 
     for i, d in enumerate(dest):
         for o in tree.query(d):
+            # STRtree query returns a list of indices for shapely >= v2.0
+            if isinstance(o, (int, np.integer)):
+                o = orig[o]
             if o.intersects(d):
-                j = idx[id(o)]
+                j = idx[hash(o.wkt)]
                 area = d.intersection(o).area
                 indicator[i, j] = area / o.area
 
@@ -177,11 +180,14 @@ def compute_intersectionmatrix(orig, dest, orig_crs=4326, dest_crs=4326):
     dest = reproject_shapes(dest, dest_crs, orig_crs)
     intersection = sp.sparse.lil_matrix((len(dest), len(orig)), dtype=float)
     tree = STRtree(orig)
-    idx = dict((id(o), i) for i, o in enumerate(orig))
+    idx = dict((hash(o.wkt), i) for i, o in enumerate(orig))
 
     for i, d in enumerate(dest):
         for o in tree.query(d):
-            j = idx[id(o)]
+            # STRtree query returns a list of indices for shapely >= v2.0
+            if isinstance(o, (int, np.integer)):
+                o = orig[o]
+            j = idx[hash(o.wkt)]
             intersection[i, j] = o.intersects(d)
 
     return intersection
@@ -466,9 +472,6 @@ def shape_availability(geometry, excluder):
         masked = ~geometry_mask(d["geometry"], shape, transform, invert=d["invert"])
         exclusions = exclusions | masked
 
-    warn(
-        "Output dtype of shape_availability changed from float to boolean.", UserWarning
-    )
     return ~exclusions, transform